+++ CHANGELOG.md
@@ -45,6 +45,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Show replies in invite screen [PR #232](https://github.com/marmot-protocol/whitenoise/pull/232)
 - Show last message for invites in chat list [#310](https://github.com/marmot-protocol/whitenoise/pull/310)
 - Create group [PR #288](https://github.com/marmot-protocol/whitenoise/pull/288)
+- Copy message to clipboard from message actions [PR #317](https://github.com/marmot-protocol/whitenoise/pull/317)
 
 ### Changed